Small thing's trying to off center an aabb via a model file. can't do it without refactor bullet some. Which is already convoluted.

+++ code/systems/entity/generic_id.cpp
@@ -5,9 +5,9 @@ namespace Generic_id {
 
 
 void
-ids_init_sequentially(size_t ids[], const size_t number_of_ids)
+ids_init_sequentially(uint32_t ids[], const size_t number_of_ids)
 {
-  for(size_t i = 0; i < number_of_ids; ++i)
+  for(uint32_t i = 0; i < number_of_ids; ++i)
   {
     ids[i] = i;
   }
